requirements.txt
channels==4.0.0
daphne==4.0.0
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
Pillow==9.4.0
django-allauth==0.53.1
ipython==8.11.0
faker==18.3.1
graphene-django==3.0.0
